refactor: extension imports page-scripts from core instead of duplicating
Replace 47 duplicated functions in extension's page-scripts.ts with
re-exports from @playwright-repl/core. Keep 11 Chrome-specific functions
(globalThis, chrome.tabs) locally in the extension.

Core changes:
- Reconcile 8 functions with extension's improvements: URL-link support
  in verifyElement/actionByRole, radio group in verifyInputValue, strict
  visible filtering in actionByText/checkByText/uncheckByText, hasText
  instead of has:getByText(exact) in role-based scoping
- Export command-scripts and missing page-scripts from index.ts
- Add ./dist/* export path for direct submodule imports

Extension page-scripts.ts: 774 → 155 lines (-80%)

Closes #855
